extract logic to validate teamWahlbezirkUndWahlID
