+++ wls-ergebnismeldung-service/src/main/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/stimmzettelerfassung/stimmzettel/StimmzettelValidator.java
@@ -3,6 +3,7 @@ package de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.stimmzett
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.exception.DataConflictException;
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.exception.ExceptionConstants;
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.stimmzettelerfassung.TeamBezirkUndWahlIDModel;
+import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.stimmzettelerfassung.TeamBezirkUndWahlIDModelValidator;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import de.muenchen.oss.wahllokalsystem.wls.common.security.domain.BezirkUndWahlID;
 import java.util.List;
@@ -15,6 +16,7 @@ import org.springframework.stereotype.Component;
 public class StimmzettelValidator {
 
   private final ExceptionFactory exceptionFactory;
+  private final TeamBezirkUndWahlIDModelValidator teamBezirkUndWahlIDModelValidator;
 
   public void validOrThrow(final List<StimmzettelOfTeamModel> listOfStimmzettel) {
     if (listOfStimmzettel == null) {
@@ -25,13 +27,11 @@ public class StimmzettelValidator {
   }
 
   public void validOrThrow(final TeamBezirkUndWahlIDModel stimmzettelOwner) {
-    if (stimmzettelOwner == null
-        || StringUtils.isBlank(stimmzettelOwner.wahlbezirkID())
-        || StringUtils.isBlank(stimmzettelOwner.wahlID())
-        || StringUtils.isBlank(stimmzettelOwner.teamID())) {
-      throw exceptionFactory.createFachlicheWlsException(
-          ExceptionConstants.STIMMZETTEL_OWNER_IDS_ARE_MISSING);
-    }
+    teamBezirkUndWahlIDModelValidator.isValidOrThrow(
+        stimmzettelOwner,
+        () ->
+            exceptionFactory.createFachlicheWlsException(
+                ExceptionConstants.STIMMZETTEL_OWNER_IDS_ARE_MISSING));
   }
 
   public void validOrThrow(final BezirkUndWahlID bezirkUndWahlID) {
