+++ wls-ergebnismeldung-service/src/test/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/stimmzettelerfassung/teamstatus/TeamErfassungStatusValidatorTest.java
@@ -1,22 +1,20 @@
 package de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.stimmzettelerfassung.teamstatus;
 
-import static org.instancio.Select.field;
-
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.exception.ExceptionConstants;
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.stimmzettelerfassung.TeamBezirkUndWahlIDModel;
+import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.stimmzettelerfassung.TeamBezirkUndWahlIDModelValidator;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.FachlicheWlsException;
+import de.muenchen.oss.wahllokalsystem.wls.common.exception.WlsException;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
-import java.util.stream.Stream;
+import java.util.function.Supplier;
 import lombok.val;
 import org.assertj.core.api.Assertions;
 import org.instancio.Instancio;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.aggregator.ArgumentsAccessor;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -27,88 +25,30 @@ class TeamErfassungStatusValidatorTest {
 
   @Mock ExceptionFactory exceptionFactory;
 
+  @Mock TeamBezirkUndWahlIDModelValidator teamBezirkUndWahlIDModelValidator;
+
   @InjectMocks TeamErfassungStatusValidator unitUnderTest;
 
   @Nested
   class IsValidOrThrow {
 
     @Nested
-    class OfWahlbezirkErfassungsteamIDModel {
+    class OfTeamBezirkUndWahlIDModel {
+
+      @Captor ArgumentCaptor<Supplier<WlsException>> captor;
 
       @Test
-      void should_notThrowAnyException_when_idIsValid() {
-        val idToValidate = Instancio.create(TeamBezirkUndWahlIDModel.class);
+      void should_useTeamBezirkUndWahlIDValidator_when_called() {
+        val id = Instancio.create(TeamBezirkUndWahlIDModel.class);
 
-        Assertions.assertThatNoException()
-            .isThrownBy(() -> unitUnderTest.isValidOrThrow(idToValidate));
-      }
+        unitUnderTest.isValidOrThrow(id);
 
-      @ParameterizedTest(name = "throw exception when {1}")
-      @MethodSource("invalidWahlbezirkErfassungsteamID")
-      void should_throwException_when_idIsInvalid(final ArgumentsAccessor arguments) {
-        val idToValidate = arguments.get(0, TeamBezirkUndWahlIDModel.class);
-        val mockedWlsException =
-            FachlicheWlsException.withCode("000").buildWithMessage("mocked wls exception");
-
-        Mockito.when(
+        Mockito.verify(teamBezirkUndWahlIDModelValidator)
+            .isValidOrThrow(Mockito.eq(id), captor.capture());
+        Assertions.assertThat(captor.getValue().get())
+            .isEqualTo(
                 exceptionFactory.createFachlicheWlsException(
-                    ExceptionConstants.STIMMZETTELERFASSUNG_TEAM_STATUS_INVALID_IDs))
-            .thenReturn(mockedWlsException);
-
-        Assertions.assertThatException()
-            .isThrownBy(() -> unitUnderTest.isValidOrThrow(idToValidate))
-            .usingRecursiveComparison()
-            .isEqualTo(mockedWlsException);
-      }
-
-      public static Stream<Arguments> invalidWahlbezirkErfassungsteamID() {
-        return Stream.of(
-            Arguments.of(null, "id is null"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::wahlID), null)
-                    .create(),
-                "wahlID is null"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::wahlID), "")
-                    .create(),
-                "wahlID is empty string"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::wahlID), "   ")
-                    .create(),
-                "wahlID is blank string"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::wahlbezirkID), null)
-                    .create(),
-                "wahlbezirkID is null"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::wahlbezirkID), "")
-                    .create(),
-                "wahlbezirkID is empty string"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::wahlbezirkID), "   ")
-                    .create(),
-                "wahlbezirkID is blank string"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::teamID), null)
-                    .create(),
-                "teamID is null"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::teamID), "")
-                    .create(),
-                "teamID is empty string"),
-            Arguments.of(
-                Instancio.of(TeamBezirkUndWahlIDModel.class)
-                    .set(field(TeamBezirkUndWahlIDModel::teamID), "   ")
-                    .create(),
-                "teamID is blank string"));
+                    ExceptionConstants.STIMMZETTELERFASSUNG_TEAM_STATUS_INVALID_IDs));
       }
     }
